Add comments to embrakes/main.hpp

+++ src/embrakes/main.hpp
@@ -64,10 +64,10 @@ class Main : public Thread
     data::StateMachine     sm_data_;
     data::EmergencyBrakes  em_brakes_;
     data::Telemetry        tlm_data_;
-    int                    command_pins_[2];
-    int                    button_pins_[2];
-    StepperInterface*      m_brake_;
-    StepperInterface*      f_brake_;
+    uint8_t                command_pins_[2];  // GPIO pin numbers for sending commands to brakes
+    uint8_t                button_pins_[2];   // GPIO pin numbers for retrieving brake status
+    StepperInterface*      m_brake_;          // Stepper for electromagnetic brakes
+    StepperInterface*      f_brake_;          // Stepper for friction brakes
 };
 
 }}
